chore: move king logic to servo instead
Since the servo board will be present on all rovers, it is suitable that
it acts as king. Otherwise we will end up with multiple configurations.
